+++ hamigaki/trunk/libs/audio/test/pcm_sink_test.cpp
@@ -24,7 +24,7 @@ namespace ut = boost::unit_test;
 
 using namespace hamigaki::audio::test;
 
-void pcm_sink_test_aux(const audio::pcm_format& fmt, double freq)
+void pcm_sink_test_aux(const audio::pcm_format& fmt, float freq)
 {
     std::ostringstream os;
     os
